Get Spotify audio analysis tidily

Fetches the Spotify audio analysis for a track using list columns rather than lists of lists.

get_tidy_audio_analysis(track_uri, ...)

Arguments

track_uri

A string with a Spotify track URI.

...

Additional parameters passed to get_track_audio_analysis.

Details

See the Spotify developer documentation for details about the information included in an audio analysis.

Examples

get_tidy_audio_analysis("6IQILcYkN2S2eSu5IHoPEH")
#> # A tibble: 1 × 13
#>   analyzer_ve…¹ durat…² end_o…³ start…⁴ tempo…⁵ time_…⁶ key_c…⁷ mode_…⁸ bars    
#>   <chr>           <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl> <list>  
#> 1 4.0.0            673.    1.14    665.   0.808   0.123   0.285   0.503 <tibble>
#> # … with 4 more variables: beats <list>, tatums <list>, sections <list>,
#> #   segments <list>, and abbreviated variable names ¹​analyzer_version,
#> #   ²​duration, ³​end_of_fade_in, ⁴​start_of_fade_out, ⁵​tempo_confidence,
#> #   ⁶​time_signature_confidence, ⁷​key_confidence, ⁸​mode_confidence