Transpose chroma vectors to C

Given a key estimate, transpose chroma vectors back to C.

compmus_c_transpose(dat, key)

Arguments

dat
Data frame with a pitches column containing chroma vectors.

key
Character string naming the global or local tonal centre.

Examples

library(tidyverse)
spotifyr::get_playlist_audio_features("", "37i9dQZF1DWWEcRhUVtL8n") %>%
  slice(1:20) %>%
  add_audio_analysis() %>%
  mutate(segments = map2(segments, key, compmus_c_transpose))

#> # A tibble: 20 x 74
#>    playlist_id  playlist_name playlist_img     playlist_owner_… playlist_owner_…
#>    <chr>        <chr>         <chr>            <chr>            <chr>           
#>  1 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#>  2 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#>  3 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#>  4 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#>  5 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#>  6 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#>  7 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#>  8 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#>  9 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 10 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 11 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 12 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 13 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 14 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 15 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 16 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 17 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 18 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 19 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> 20 37i9dQZF1DW… Indie Pop     https://i.scdn.… Spotify          spotify         
#> # … with 69 more variables: danceability <dbl>, energy <dbl>, key <int>,
#> #   loudness <dbl>, mode <int>, speechiness <dbl>, acousticness <dbl>,
#> #   instrumentalness <dbl>, liveness <dbl>, valence <dbl>, tempo <dbl>,
#> #   track.id <chr>, analysis_url <chr>, time_signature <int>, added_at <chr>,
#> #   is_local <lgl>, primary_color <lgl>, added_by.href <chr>,
#> #   added_by.id <chr>, added_by.type <chr>, added_by.uri <chr>,
#> #   added_by.external_urls.spotify <chr>, track.artists <list>,
#> #   track.available_markets <list>, track.disc_number <int>,
#> #   track.duration_ms <int>, track.episode <lgl>, track.explicit <lgl>,
#> #   track.href <chr>, track.is_local <lgl>, track.name <chr>,
#> #   track.popularity <int>, track.preview_url <chr>, track.track <lgl>,
#> #   track.track_number <int>, track.type <chr>, track.uri <chr>,
#> #   track.album.album_type <chr>, track.album.artists <list>,
#> #   track.album.available_markets <list>, track.album.href <chr>,
#> #   track.album.id <chr>, track.album.images <list>, track.album.name <chr>,
#> #   track.album.release_date <chr>, track.album.release_date_precision <chr>,
#> #   track.album.total_tracks <int>, track.album.type <chr>,
#> #   track.album.uri <chr>, track.album.external_urls.spotify <chr>,
#> #   track.external_ids.isrc <chr>, track.external_urls.spotify <chr>,
#> #   video_thumbnail.url <lgl>, key_name <chr>, mode_name <chr>, key_mode <chr>,
#> #   analyzer_version <chr>, duration <dbl>, end_of_fade_in <dbl>,
#> #   start_of_fade_out <dbl>, tempo_confidence <dbl>,
#> #   time_signature_confidence <dbl>, key_confidence <dbl>,
#> #   mode_confidence <dbl>, bars <list>, beats <list>, tatums <list>,
#> #   sections <list>, segments <list>