Gather chroma vectors

These functions gather chroma or timbre vectors into long format for Computational Musicology, as generated by get_tidy_audio_analysis.

compmus_gather_chroma(dat)

compmus_gather_timbre(dat)

Arguments

dat
Data frame with a pitches or timbre column.

Functions

compmus_gather_timbre: Gather timbre vectors

Examples

library(tidyverse)
get_tidy_audio_analysis('6IQILcYkN2S2eSu5IHoPEH') %>%
    select(segments) %>% unnest(segments) %>%
    compmus_gather_chroma

#> # A tibble: 52,656 x 9
#>    timbre start duration confidence loudness_start loudness_max_ti… loudness_max
#>    <list> <dbl>    <dbl>      <dbl>          <dbl>            <dbl>        <dbl>
#>  1 <dbl …  0       1.14           0          -60            0              -60  
#>  2 <dbl …  1.14    0.255          1          -60            0.00844        -35.5
#>  3 <dbl …  1.40    0.337          1          -54.3          0.0432         -34.9
#>  4 <dbl …  1.74    0.255          1          -53.7          0.00787        -33.9
#>  5 <dbl …  1.99    0.291          1          -53.1          0.0436         -33.5
#>  6 <dbl …  2.28    0.296          1          -54.0          0.0491         -34.1
#>  7 <dbl …  2.58    0.296          1          -53.3          0.0493         -33.6
#>  8 <dbl …  2.87    0.296          1          -53.5          0.0501         -34.3
#>  9 <dbl …  3.17    0.296          1          -53.8          0.0471         -34.6
#> 10 <dbl …  3.47    0.291          1          -54.0          0.0433         -34.2
#> # … with 52,646 more rows, and 2 more variables: pitch_class <fct>, value <dbl>

get_tidy_audio_analysis('5ZLkc5RY1NM4FtGWEd6HOE') %>%
    select(segments) %>% unnest(segments) %>%
    compmus_gather_timbre

#> # A tibble: 9,324 x 9
#>    pitches start duration confidence loudness_start loudness_max_ti…
#>    <list>  <dbl>    <dbl>      <dbl>          <dbl>            <dbl>
#>  1 <dbl [… 0       0.407       0             -60              0     
#>  2 <dbl [… 0.407   0.0633      1             -60              0.0308
#>  3 <dbl [… 0.470   0.639       0.706         -12.1            0.211 
#>  4 <dbl [… 1.11    0.236       0.166         -11.4            0.0455
#>  5 <dbl [… 1.34    0.340       0.872         -11.6            0.0799
#>  6 <dbl [… 1.68    0.423       0.672          -9.78           0.0878
#>  7 <dbl [… 2.11    0.402       0.548          -9.52           0.132 
#>  8 <dbl [… 2.51    0.347       0.091         -11.7            0.0307
#>  9 <dbl [… 2.86    0.186       0.724         -14.6            0.0524
#> 10 <dbl [… 3.04    0.0924      0.134          -9.00           0.0584
#> # … with 9,314 more rows, and 3 more variables: loudness_max <dbl>,
#> #   basis <chr>, value <dbl>