Normalise vectors

We use a number of vector normalisation strategies in Computational Musicology. This function brings them together into one place, along with common alternative names.

compmus_normalise(v, method = "euclidean")

compmus_normalize(v, method = "euclidean")

Arguments

v
A numeric vector.

method
A character string indicating which normalization to use (see Details). Default is the Euclidean norm.

Details

The following methods are supported.

identity,id

No normalisation.

harmonic

Harmonic mean.

manhattan,L1

Manhattan (L1) norm.

euclidean,L2

Euclidean (L2) norm.

chebyshev,maximum

Chebyshev (maximum) norm.

aitchison,clr

Aitchison's clr transformation.

softmax

Softmax.

Functions

compmus_normalize: Normalize vectors

Examples

library(tidyverse)
get_tidy_audio_analysis("6IQILcYkN2S2eSu5IHoPEH") %>%
  select(segments) %>%
  unnest(segments) %>%
  mutate(pitches = map(pitches, compmus_normalise, "euclidean"))

#> # A tibble: 4,388 x 9
#>    start duration confidence loudness_start loudness_max_ti… loudness_max
#>    <dbl>    <dbl>      <dbl>          <dbl>            <dbl>        <dbl>
#>  1  0       1.14           0          -60            0              -60  
#>  2  1.14    0.255          1          -60            0.00844        -35.5
#>  3  1.40    0.337          1          -54.3          0.0432         -34.9
#>  4  1.74    0.255          1          -53.7          0.00787        -33.9
#>  5  1.99    0.291          1          -53.1          0.0436         -33.5
#>  6  2.28    0.296          1          -54.0          0.0491         -34.1
#>  7  2.58    0.296          1          -53.3          0.0493         -33.6
#>  8  2.87    0.296          1          -53.5          0.0501         -34.3
#>  9  3.17    0.296          1          -53.8          0.0471         -34.6
#> 10  3.47    0.291          1          -54.0          0.0433         -34.2
#> # … with 4,378 more rows, and 3 more variables: loudness_end <dbl>,
#> #   pitches <list>, timbre <list>