Get Spotify audio analysis tidily

Fetches the Spotify audio analysis for a track using list columns rather than lists of lists.

get_tidy_audio_analysis(track_uri, ...)

Arguments

track_uri
A string with a Spotify track URI.

...
Additional parameters passed to get_track_audio_analysis.

Details

See the Spotify developer documentation for details about the information included in an audio analysis.

Examples

get_tidy_audio_analysis('6IQILcYkN2S2eSu5IHoPEH')

#> # A tibble: 1 x 13
#>   analyzer_version duration end_of_fade_in start_of_fade_o… tempo_confidence
#>   <chr>               <dbl>          <dbl>            <dbl>            <dbl>
#> 1 4.0.0                673.           1.14             665.            0.808
#> # … with 8 more variables: time_signature_confidence <dbl>,
#> #   key_confidence <dbl>, mode_confidence <dbl>, bars <list>, beats <list>,
#> #   tatums <list>, sections <list>, segments <list>